ai-feedback-video2/ai-feedback-video2/backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
azure-cognitiveservices-speech==1.47.0
azure-core==1.36.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.9.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
